+++ README.md
@@ -1,0 +1,3 @@
+# Dynamic learning GraphQL server
+
+Repository for the back end of dynamic learning.
