+++ src/models/workbookFolders.ts
@@ -1,0 +1,20 @@
+import mongoose from "mongoose";
+
+const Schema = mongoose.Schema;
+
+const WorkbookFolderSchema = new Schema({
+  title: {
+    type: String,
+    required: true,
+  },
+  owner: {
+    type: String,
+    required: true,
+  },
+  parentId: {
+    type: Schema.Types.ObjectId,
+    ref: "workbookfolders",
+  },
+});
+
+export default mongoose.model("Workbookfolder", WorkbookFolderSchema);
